pyproject.toml
[project]
name = "david"
version = "v0.1.0"
description = "Django template"
readme = "README.md"
requires-python = ">=3.14.0,<3.15.0"
dependencies = [
    "celery[librabbitmq]==5.6.0",
    "django==5.2.7",
    "django-celery-beat==2.8.1",
    "django-celery-results==2.6.0",
    "django-command-log==2.2.1",
    "django-redis==6.0.0",
    "djangorestframework==3.16.1",
    "drf-spectacular==0.27.0",
    "gunicorn==23.0.0",
    "ipython>=9.6.0",
    "psycopg2-binary==2.9.11",
    "pytz==2025.2",
    "redis==7.1.0",
    "requests==2.32.5",
    "ruff>=0.14.8",
    "uvicorn[standard]==0.38.0",
]

[project.optional-dependencies]
dev = [
    "factory-boy>=3.3.3",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "pytest-factoryboy>=2.8.1",
]

[tool.setuptools]
py-modules = []

[tool.pytest.ini_options]
log_cli = true
log_cli_format = "%(levelname)s %(asctime)s %(message)s"
django_find_project = true
addopts = '-p no:warnings --color=yes --ds=david.settings.test'
norecursedirs = '''
.git
.github
.pytest_cache
.ruff_cache
'''

[tool.bandit]
exclude_dirs = [
    "david/settings/",
    "david/tests/factories.py"
]
skips = ["B101", "B308", "B703"]

[tool.black]
line-length = 120
target_version = ['py314']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.idea
  | \.vscode
  | \.pytest_cache
  | \.ruff_cache
  | \.venv
)/
'''

[tool.curlylint.rules]
indent = 2
html_has_lang = true

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py314"
exclude = [
    "david/settings/"
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # Pyflakes
    "N",  # pep8-naming
    "UP",  # pyupgrade
    "B",  # flake8-bugbear
    "S",  # flake8-bandit
    "BLE",  # flake8-blind-except
    "A",  # flake8-builtins
    "DJ",  # flake8-django
    "I",  # isort
    "INP",  # flake8-no-pep420
    "RUF"  # Ruff-specific rules
]
exclude = [
    ".git",
    ".idea",
    ".vscode",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "venv",
]
ignore = [
    "DJ001", # Null values for model char fields
    "ERA001",  # Disable eradicate (commented code removal)
    # Disable Conflicting lint rules,
    # see https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "W191",
    "E501",
    "E111",
    "E117",
    "D206",
    "D300",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "COM812",
    "COM819",
    "ISC001",
    "ISC002",
    "S603",
    # Disable unused `noqa` directive
    "RUF100",
    "RUF012",  # Mutable class attributes should be annotated with `typing.ClassVar`
    "N806",  # Variable `CONSTANT_NAME` in function should be lowercase
]

[tool.ruff.lint.pycodestyle]
ignore-overlong-task-comments = true

[tool.ruff.lint.isort]
section-order = ["future", "standard-library", "third-party", "first-party", "django", "david", "local-folder"]
lines-after-imports = 2

[tool.ruff.lint.isort.sections]
"django" = ["django", "apscheduler", "django_apscheduler", "django_rq", "command_log", "rest_framework"]
"david" = ["david"]

[tool.ruff.lint.per-file-ignores]
# Ignore "E402", "F403", "F405" (import violations) in __init__.py files.
# Ignore "S" (flake8-bandit) and "N802" (function name should be lowercase) in tests and docs.
# Ignore "RUF" (Ruff-specific rules) and "I" (isort) in migrations.
"__init__.py" = ["E402", "F403", "F405"]
"**/{tests,docs}/*" = ["E402", "F403", "F405", "S", "N802", "N815"]
"**/*test*.py" = ["E402", "F403", "F405", "S", "N802"]
"**/{settings}/*" = ["E402", "F403", "F405"]
"**/migrations/*" = ["RUF", "I"]
